==> 1_Data_Preparation.ipynb <==
import pandas as pd
import yfinance as yf

stock_list = pd.read_csv('raw/sp500_companies.csv')['Symbol']

start = '2014-01-01'
end = '2020-03-31'
max_length = 0
stock_dict = {}
for stock in stock_list:
    stock_dict[stock] = yf.download(tickers=stock, start=start, end=end, interval='1d').dropna()
    if stock_dict[stock].shape[0] > max_length:
        max_length = stock_dict[stock].shape[0]

for i, stock in enumerate(stock_list):
    if stock_dict[stock].shape[0] < max_length: # remove the stock if the stock does not have the full history
        stock_list.drop(i, inplace = True)
    else:
        stock_dict[stock].to_csv('data/stock/{}.csv'.format(stock))
stock_list.to_csv('data/stock_list.csv', index=False)
