==> 4_Trade_Analysis.ipynb <==
import datetime
import pandas as pd
import statsmodels.api as sm

perf_df = pd.read_csv('results/trade_record.csv', index_col=0, parse_dates=['Start Date', 'End Date'])
perf_df

perf_df['Average Return'] = perf_df['Spread Return'] / ((perf_df['End Date'] - perf_df['Start Date']) / pd.Timedelta(days=1))

for i, row in enumerate(perf_df.values):
    if row[1] == datetime.datetime(2018, 12, 28):
        perf_df = perf_df.drop(i)
perf_df

perf_df = perf_df.drop('Start Date', axis=1)
perf_df = perf_df.drop('End Date', axis=1)
perf_df = perf_df.drop('a', axis=1)
perf_df = perf_df.drop('b', axis=1)
perf_df = perf_df.drop('Spread Return', axis=1)
perf_df = perf_df.drop('Position', axis=1)
perf_df

Y = perf_df.iloc[:, -1].T
X = perf_df.iloc[:, :-1]

model = sm.OLS(Y,X)
results = model.fit()

results.summary()

X = X.drop('ADF p-value', axis=1)
model = sm.OLS(Y,X)
results = model.fit()
results.summary()

X = X.drop('Correlation', axis=1)
model = sm.OLS(Y,X)
results = model.fit()
results.summary()

X = X.drop('sigma2', axis=1)
model = sm.OLS(Y,X)
results = model.fit()
results.summary()

X = X.drop('ADF t-statistic', axis=1)
model = sm.OLS(Y,X)
results = model.fit()
results.summary()

results.params.to_csv('results/OLS_params.csv', header=None)
